rf-board-organizer/requirements.txt
blinker==1.9.0
click==8.2.1
ezdxf==1.4.2
Flask==3.1.1
flask-cors==6.0.0
Flask-SQLAlchemy==3.1.1
fonttools==4.58.4
greenlet==3.2.3
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
numpy==2.3.0
numpy-stl==3.2.0
pyparsing==3.2.3
python-utils==3.9.1
SQLAlchemy==2.0.41
typing_extensions==4.14.0
Werkzeug==3.1.3
